PAR-8 Add useMovies hook for fetching movie data
